Move zxbasm into its own folder
This will unclutter the root directory
