+++ arch/zx48k/optimizer/memcell.py
@@ -6,7 +6,7 @@ from . import helpers
 from .. import backend
 from .asm import Asm
 from api.utils import flatten_list
-import asmlex
+from zxbasm import asmlex
 
 
 class MemCell(object):
