+++ parsetab/tabs.dbm.bak
@@ -1,4 +1,4 @@
 'zxbppparse', (0, 68810)
-'asmparse', (69120, 250081)
-'zxnext_asmparse', (319488, 280962)
-'zxbparser', (600576, 707013)
+'asmparse', (69120, 250690)
+'zxnext_asmparse', (320000, 281648)
+'zxbparser', (602112, 707013)
